+++ packages/assets/scripts/figma-config.js
@@ -125,6 +125,11 @@ export const v2LogosConfig = createFigmaExportConfig({
 // v3 outline icons configuration
 export const v3OutlineIconsConfig = createFigmaExportConfig({
   additionalOutputters: [
+    outputComponentsToBundle({
+      exportName: 'outlineIcons',
+      getBasename: ({ basename = '' }) => basename.replace('name=', ''),
+      output: './src/brand2025-outline-icons.ts',
+    }),
     outputComponentsToCodeConnect({
       fileId: FIGMA_CONFIG.FIGMA_FILE_ID_ICONS_V3,
       getBasename: ({ basename = '' }) => basename.replace('name=', ''),
@@ -140,6 +145,11 @@ export const v3OutlineIconsConfig = createFigmaExportConfig({
 // v3 filled icons configuration
 export const v3FilledIconsConfig = createFigmaExportConfig({
   additionalOutputters: [
+    outputComponentsToBundle({
+      exportName: 'filledIcons',
+      getBasename: ({ basename = '' }) => basename.replace('name=', ''),
+      output: './src/brand2025-filled-icons.ts',
+    }),
     outputComponentsToCodeConnect({
       fileId: FIGMA_CONFIG.FIGMA_FILE_ID_ICONS_V3,
       getBasename: ({ basename = '' }) => basename.replace('name=', ''),
@@ -181,4 +191,4 @@ export const CONFIG_FOR_ALL = [
  * Default configuration to use.
  * Set this to CONFIG_FOR_ALL to export all icons
  */
-export const DEFAULT_CONFIG = CONFIG_FOR_V2;
+export const DEFAULT_CONFIG = CONFIG_FOR_ALL;